requirements.txt
# --- web framework ---
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.3
python-multipart==0.0.20

# --- google/vertex ai ---
google-cloud-aiplatform==1.71.1
google-cloud-storage<3.0.0
google-cloud-firestore==2.19.0
vertexai==1.71.1   

# --- utilities ---
pydantic==2.11.9
typing-extensions==4.15.0
python-dotenv==1.1.1
requests>=2.31.0,<3
anyio==4.10.0
h11==0.16.0
httptools==0.6.4
watchfiles==1.1.0
websockets==15.0.1
PyYAML==6.0.2